Made BSQ 7% faster by avoiding max in find_u32_arr_larger_with_pos
Before this change, we were doing something stupid and in every loop
we would find the largest value out of the ones we had present before
comparing it to the current largest value. We can instead simply
compare every value to the current largest value and see if at least
one of them is larger before scanning them to find which specifically
is the largest of themselves.

Note that it doesn't really matter that the interior of the if
conditional with _mm_movemask_epi8 is particularly slow - it should
generally be executed far more rarely than the other case (in practice
it doesn't even show up in perf traces for me)