Made BSQ ~6-7% faster by optimizing check_square for CPU pipelines
Basically, we would generally expect that the CPU currently can't
execute more than one iteration of check_square at a time, with
pipelining/out-of-order execution, because each iteration directly
depends on the final value of the previous iteration (in particular
when it gets stored into square_size_values). Help out the CPU here by
making it so that most of the time we compute "one ahead", that is, we
make it so that most of the time we use an alternate
check_square_one_ahead function that only depends on the iteration
done two squares ago, rather than the directly previous one, so that
the CPU can hopefully execute two iterations at the same time with
pipelining/out-of-order execution.