Made BSQ 9-14% faster by using memset to handle 'o' size storing
In check_square, we currently set square_size_values[x] = 0 when we
encounter a square with a value of 'o'. It looks like it's faster to
instead do nothing in this case and have it be handled implicitly by
memset-ing all of square_size_values at the beginning of check_line.